+++ Assets/Scripts/Player/PlayerHealth.cs
@@ -2,8 +2,8 @@ using UnityEngine;
 
 public class PlayerHealth : MonoBehaviour
 {
-    [SerializeField] private float totalHealth = 10;
-    private float health;
+    [SerializeField] private int totalHealth = 10;
+    private int health;
 
     void Update(){
         Debug.Log($"Health is {health}");
@@ -12,7 +12,7 @@ public class PlayerHealth : MonoBehaviour
         health=totalHealth;
     }
 
-    private void UpdateHealth(float dmg) {
+    private void UpdateHealth(int dmg) {
         health -= dmg;
     }
 
@@ -21,7 +21,7 @@ public class PlayerHealth : MonoBehaviour
     }
 
     public void FullHeal(){
-        float tempHealth = totalHealth-health;
+        int tempHealth = totalHealth-health;
 
         health+=tempHealth;
     }
